Divide project into core and dom modules
