+++ domino-core/shared/src/main/scala/org/domino/AttributeFactory.scala
@@ -1,0 +1,5 @@
+package org.domino
+
+case class AttributeFactory[T, A](f: T => A) {
+  def :=(value: T): A = f(value)
+}
