+++ domino-core/shared/src/main/scala/org/domino/Attributes.scala
@@ -1,4 +1,4 @@
-package org.domino.html
+package org.domino
 
 trait Attributes {
 
@@ -68,69 +68,6 @@ trait Attributes {
   val muted = AttributeFactory(Muted)
   val name = AttributeFactory(Name)
   val novalidate = AttributeFactory(NoValidate)
-  val onAbort = AttributeFactory(OnAbort)
-  val onActivate = AttributeFactory(OnActivate)
-  val onBeforeCut = AttributeFactory(OnBeforeCut)
-  val onBeforeActivate = AttributeFactory(OnBeforeActivate)
-  val onBeforeDeactivate = AttributeFactory(OnBeforeDeactivate)
-  val onBeforeCopy = AttributeFactory(OnBeforeCopy)
-  val onBeforePaste = AttributeFactory(OnBeforePaste)
-  val onBlur = AttributeFactory(OnBlur)
-  val onCanPlay = AttributeFactory(OnCanPlay)
-  val onCanPlayThrough = AttributeFactory(OnCanPlayThrough)
-  val onChange = AttributeFactory(OnChange)
-  val onClick = AttributeFactory(OnClick)
-  val onContextMenu = AttributeFactory(OnContextMenu)
-  val onCueChange = AttributeFactory(OnCueChange)
-  val onDeactivate = AttributeFactory(OnDeactivate)
-  val onDoubleClick = AttributeFactory(OnDoubleClick)
-  val onDrag = AttributeFactory(OnDrag)
-  val onDragEnd = AttributeFactory(OnDragEnd)
-  val onDragEnter = AttributeFactory(OnDragEnter)
-  val onDragLeave = AttributeFactory(OnDragLeave)
-  val onDragOver = AttributeFactory(OnDragOver)
-  val onDragStart = AttributeFactory(OnDragStart)
-  val onDrop = AttributeFactory(OnDrop)
-  val onEmptied = AttributeFactory(OnEmptied)
-  val onEnded = AttributeFactory(OnEnded)
-  val onFocus = AttributeFactory(OnFocus)
-  val onFocusIn = AttributeFactory(OnFocusIn)
-  val onFocusOut = AttributeFactory(OnFocusOut)
-  val onHelp = AttributeFactory(OnHelp)
-  val onInput = AttributeFactory(OnInput)
-  val onKeyDown = AttributeFactory(OnKeyDown)
-  val onKeyPress = AttributeFactory(OnKeyPress)
-  val onKeyUp = AttributeFactory(OnKeyUp)
-  val onLoad = AttributeFactory(OnLoad)
-  val onLoadedData = AttributeFactory(OnLoadedData)
-  val onLoadedMetadata = AttributeFactory(OnLoadedMetadata)
-  val onLoadStart = AttributeFactory(OnLoadStart)
-  val onMouseDown = AttributeFactory(OnMouseDown)
-  val onMouseEnter = AttributeFactory(OnMouseEnter)
-  val onMouseLeave = AttributeFactory(OnMouseLeave)
-  val onMouseMove = AttributeFactory(OnMouseMove)
-  val onMouseOut = AttributeFactory(OnMouseOut)
-  val onMouseOver = AttributeFactory(OnMouseOver)
-  val onMouseUp = AttributeFactory(OnMouseUp)
-  val onMouseWheel = AttributeFactory(OnMouseWheel)
-  val onPause = AttributeFactory(OnPause)
-  val onPlay = AttributeFactory(OnPlay)
-  val onPlaying = AttributeFactory(OnPlaying)
-  val onProgress = AttributeFactory(OnProgress)
-  val onRateChange = AttributeFactory(OnRateChange)
-  val onReadyStateChange = AttributeFactory(OnReadyStateChange)
-  val onReset = AttributeFactory(OnReset)
-  val onScroll = AttributeFactory(OnScroll)
-  val onSeeked = AttributeFactory(OnSeeked)
-  val onSeeking = AttributeFactory(OnSeeking)
-  val onSelectStart = AttributeFactory(OnSelectStart)
-  val onSelect = AttributeFactory(OnSelect)
-  val onStalled = AttributeFactory(OnStalled)
-  val onSubmit = AttributeFactory(OnSubmit)
-  val onSuspend = AttributeFactory(OnSuspend)
-  val onTimeUpdate = AttributeFactory(OnTimeUpdate)
-  val onVolumeChange = AttributeFactory(OnVolumeChange)
-  val onWaiting = AttributeFactory(OnWaiting)
   val open = AttributeFactory(Open)
   val optimum = AttributeFactory(Optimum)
   val pattern = AttributeFactory(Pattern)
@@ -174,9 +111,4 @@ trait Attributes {
   val wrap = AttributeFactory(Wrap)
 
   def data(name: String) = AttributeFactory((t: String) => CustomData(name, t))
-
-  case class AttributeFactory[T, A](f: T => A) {
-    def :=(value: T): A = f(value)
-  }
-
 }
