+++ domino-core/shared/src/main/scala/org/domino/Element.scala
@@ -1,4 +1,4 @@
-package org.domino.html
+package org.domino
 
 sealed trait Node {
   def render: String
