--- src/test/scala/org/domino/UnitTest.scala
@@ -1,5 +1,0 @@
-package org.domino
-
-import org.scalatest.{BeforeAndAfter, FunSuite, Matchers}
-
-abstract class UnitTest extends FunSuite with Matchers with BeforeAndAfter
